refactor(game): split scene3d RNG + orbital mechanics → scene3d/rng.ts + orbital.ts (logic sama)
