+++ apps/game/src/renderer/scene3d/index.ts
@@ -40,44 +40,8 @@ export interface Scene3D {
   dispose(): void;
 }
 
-// ---------------------------------------------------------------------------
-// Deterministic RNG & orbit math (client-side — sama deterministik server §2.3)
-// ---------------------------------------------------------------------------
-function mulberry32(seed: number): () => number {
-  let a = seed >>> 0;
-  return () => {
-    a |= 0; a = (a + 0x6d2b79f5) | 0;
-    let t = Math.imul(a ^ (a >>> 15), 1 | a);
-    t = (t + Math.imul(t ^ (t >>> 7), 61 | t)) ^ t;
-    return ((t ^ (t >>> 14)) >>> 0) / 4294967296;
-  };
-}
-
-interface OrbitSpec {
-  /** semi-major axis (render units). */
-  semimajor: number;
-  /** eccentricity 0..<1. */
-  eccentricity: number;
-  /** rotation per tick (rad). */
-  omega: number;
-  /** phase offset (rad). */
-  phase: number;
-  inclination: number;
-}
-
-function keplerPosition(o: OrbitSpec, tick: number): THREE.Vector3 {
-  const M = o.phase + tick * o.omega;
-  const e = o.eccentricity;
-  // Iterasi 3× buat Mean→Eccentric anomaly (cukup buat e<0.4; deterministik).
-  let E = M;
-  for (let i = 0; i < 3; i++) E = M + e * Math.sin(E);
-  const x = o.semimajor * (Math.cos(E) - e);
-  const z = o.semimajor * Math.sqrt(1 - e * e) * Math.sin(E);
-  // Rossi ke bidang miring (inclination).
-  const ci = Math.cos(o.inclination);
-  const y = z * Math.sin(o.inclination);
-  return new THREE.Vector3(x, y, z * ci);
-}
+import { mulberry32 } from "./rng";
+import { keplerPosition, type OrbitSpec } from "./orbital";
 
 // ---------------------------------------------------------------------------
 // Body catalogue (blueprint §2.1/§2.2) — tipe berbeda, bukan bolak warna.
